\name{testOptim}

\alias{testOptim}

\title{
  
  Test Optimization Speeds
  
}

\description{

  The function \code{testOptim} estimates a Weibull event count
  model where the dependent variable is the number of agencies created
  during a presidency.
}

\usage{
testOptim()
}

\details{
  Describe the model.

}

\value{
NULL
}

\author{

  Jonathan Paul Olmsted, \email{jpolmsted@NOSPAM.gmail.com},
  \url{www.rochester.edu/college/gradstudents/jolmsted/}
  
}

\note{

  This function was designed to finish evaluation in just over 2 minutes
  on the author's machine.

}

\examples{
## Code Not Run
## system.time(testOptim())
}

