\name{testMASS}
\alias{testMASS}

\title{
Execute a mass of statistical methods from MASS for timing.
}

\description{This function runs a varied collection of statistical
methods per the \code{MASS} package. Its purpose is to fills space with a
realistic set of tasks.}

\usage{
testMASS()
}

\details{
  add stuff
}

\value{

}
\references{
add link
}
\author{
J. P. Olmsted <jpolmsted@gmail.com>
}
\note{
%%  ~~further notes~~
}

\section{Warning}{Evaluation can take a few minutes!}


\seealso{
\code{\link{MASS}}
}

\examples{
\dontrun{system.time(testMASS())}
  }

\keyword{utilities}

